moving run_template to _jobs.py
